+++ scripts/cleanup.sh
@@ -1,0 +1,6 @@
+cd packages/dart_mappable
+cp pubspec.yaml.save pubspec.yaml
+rm pubspec.yaml.save
+cd ../dart_mappable_builder
+cp pubspec.yaml.save pubspec.yaml
+rm pubspec.yaml.save